(gauge_callback): fix -Wconversion warnings.

+++ lib/widget/gauge.c
@@ -80,8 +80,8 @@ gauge_callback (Widget * w, Widget * sender, widget_msg_t msg, int parm, void *d
         {
             int gauge_len;
             int percentage, columns;
-            long total = g->max;
-            long done = g->current;
+            int total = g->max;
+            int done = g->current;
 
             if (total <= 0 || done < 0)
             {
